finish binding

+++ movement/include/imu_movement.hpp
@@ -39,7 +39,7 @@ void gyro_drive_straight(int from_speed, int to_speed, std::function<bool()> sto
  * @param accel_per_sec how fast to accelerate, defaults to 500
  * @param updates_per_sec how many updates the function will do per sec, defaults to 200
  */
-void gyro_turn_degrees(Speed from_speed, Speed to_speed, int degrees, double accel_per_sec = get_config().getDouble("gyro_linear_accel"), int updates_per_sec = get_config().getInt("gyro_updates_per_sec"));
+void gyro_turn_degrees(Speed from_speed, Speed to_speed, double degrees, double accel_per_sec = get_config().getDouble("gyro_linear_accel"), int updates_per_sec = get_config().getInt("gyro_updates_per_sec"));
 
 /**
  * @brief Turns the create, starting at rest and ending at rest, and turning at max at max_speed
@@ -50,7 +50,7 @@ void gyro_turn_degrees(Speed from_speed, Speed to_speed, int degrees, double acc
  * @param accel_per_sec how fast to accelerate, defaults to 500
  * @param updates_per_sec how many updates the function will do per sec, defaults to 200
  */
-void gyro_turn_degrees_v2(int max_speed, int degrees, int min_speed = get_config().getInt("gyro_turn_v2_min_speed"), double accel_per_sec = get_config().getDouble("gyro_linear_accel"), int updates_per_sec = get_config().getInt("gyro_updates_per_sec"));
+void gyro_turn_degrees_v2(int max_speed, double degrees, int min_speed = get_config().getInt("gyro_turn_v2_min_speed"), double accel_per_sec = get_config().getDouble("gyro_linear_accel"), int updates_per_sec = get_config().getInt("gyro_updates_per_sec"));
 
 #endif
 /**@}*/
